pyproject.toml
[project]
name = "skill-runner"
version = "0.2.0"
description = "Agent Skill Runner"
authors = [
    {name = "Joshua Reed", email = "leike0813@gmail.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "apscheduler>=3.10.0",
    "jinja2>=3.1.0",
    "pyyaml",
    "jsonschema",
    "tomlkit",
    "yacs>=0.1.8",
    "python-multipart",
    "aiosqlite>=0.20.0",
    "pywinpty>=2.0.15; platform_system == 'Windows'",
]
requires-python = ">=3.11"
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "httpx",
]

[project.scripts]
agent-harness = "agent_harness.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
[tool.setuptools.packages.find]
where = ["."]
include = ["server*", "e2e_client*", "agent_harness*"]
exclude = ["data*", "skills*", "tests*", "references*"]

[tool.pytest.ini_options]
markers = [
    "e2e: REST-level end-to-end tests",
]
